Update default version of Tempest
Previously, refstack-client checked out version 7c8fcc67 of Tempest
by default.  As of the 2015.05 DefCore Guideline, however, a few
flags are being dropped due to bugfixes in Tempest [1][2].  In order
to ensure that users get a verison of Tempest that includes those
fixes, it makes sense to update our default to a version recent
enough to include them.  The two bugfixes I'm aware of are both
present as of 98f62ee, so this patch updates refstack-client
to use it by default.

[1] https://review.openstack.org/#/c/189931/
[2] https://review.openstack.org/#/c/189978/

Change-Id: I639a8974a15bb0bde24432ac8cfd193207071ab9