Update default version of Tempest
Previously, refstack-client checked out version 98f62ee of Tempest
by default.  As of the 2015.07 (planned) DefCore Guideline, however,
a few flags are being dropped due to changes in Tempest [1] that
make SSH tests more palletable.  In order to ensure that users get a
verison of Tempest that includes those fixes, it makes sense to update
our default to a version recent enough to include them.  The last
patchs for the new SSH auth strategy blueprint [2] landed in commit
ffe09dd [3], so this patch updates refstack-client to use it by default.

[1] https://review.openstack.org/#/c/189927/
[2]
https://blueprints.launchpad.net/openstack/?searchtext=ssh-auth-strategy
[3] https://review.openstack.org/#/c/165504/

Change-Id: Iac6d7de42d7a035a8bd171d6ab5dc64d7300dca2