Update default Tempest version to SHA 551e1a97(Sept 11, 2015).
Although refstack-client does install a default Tempest check
out version (the current version is SHA [1]),  refstack-client
has also offered options for user to install other versions of
Tempest if needed.  However, the "test_index_no_params" test in
Tempest was recently renamed to "test_list_no_params" in patch
[2] which was merged on Sept 16, 2015.  As such, test results
collected with any version of Tempest that's later than [3] will
never be able to obtain a pass status for the
"test_index_no_params" test case which is included in the Board
approved 2015.07 guideline.  For tests run against 2015.07, it
is important that users use a version of Tempest that doesn't
include the name change.

In order to allow for users to get a verison of Tempest that
includes recent Tempest fixes, refStack-client will be updated
its default Tempest version to [4] which is the most recent
version that doesn't include the name change.

More information about the test name change discussion can be
found in [5].

[1] ffe09dd61170bfccf7d60ef9c8eba2883ab6c148 (July 16, 2015)
[2] https://review.openstack.org/#/c/223796/
[3] eb9941ba06d87b3fb458ce472cae4c594ad8c0cf
[4] 551e1a9701e7e2b3edf6d49a2eaa62b7ab2435ad (September 11, 22015)
[5] https://review.openstack.org/#/c/229177/

Change-Id: I34e73a20e8129ac7e8bcd2a6b47c1b3591b37838