Call pip by executing python -m pip to avoid issue with long shebang
If refstack-client is cloned to a place in the filesystem with
a pretty long path, such that the path to pip in the virtualenv
exceeds about 128 characters (with the #! in the shebang line)
then setup_env will fail because the path to the interpreter
gets truncated. This is a known issue with pip due to a kernel
limitation on the size of the shebang line.

Change-Id: Ie2cce5647960c45032c03f008501b8f87c0991cd
Closes-Bug: #1643046