Change auth_url setup according to Tempest docs
Refstack-client fails when tempest value "api_v3" under section
identity_feature_group is not set. But according to Tempest docs:
https://docs.openstack.org/tempest/latest/configuration.html#keystone-connection-info
api_v3 is not required to stablish a connection with keystone.
Furthermore it defaults to True which will enable v3 features
unless explicitly set to False.

This fix simplifies the way we set auth_url

Change-Id: I1a4ea9cdcc625f94f978e1a929de1f543000051d