Fix generate-tempestconf-file role
The role has been using python-tempestconf master branch when
gating refstack-client or python-tempestconf patches.
This patch adds a new argument to setup_env script which can
be used to specify a path to python-tempestconf source code.
This argument is used, when running the setup_env script from
the role, to point the script to the zuul source of
python-tempestconf code which contains the newest commits.

Change-Id: I65f9043a7bb0e6f045b6442679dd99dbd6aeb2de