Check python version when installing deps via yum
It seems that newer yum will not install all packages provided if one of
them is not found.
setup_env script installs for example python-devel and python3-devel
where one of them is for installation with python2 and the other
for python3.
In order to avoid situations when required packages are not installed
because one of the packages is not available, the patch adds logic
which will install deps based on the python version.

Change-Id: Ib119b283fac3cda340813f48b10c0a3cc4573891